+++ instrument-panel/instruments/adi.cpp
@@ -174,7 +174,7 @@ void adi::render()
 void adi::update()
 {
     // Check for position or size change
-    long *settings = globals.simVars->readSettings(name, xPos, yPos, size);
+    int *settings = globals.simVars->readSettings(name, xPos, yPos, size);
 
     xPos = settings[0];
     yPos = settings[1];
