+++ instrument-panel/instruments/spitfire/vsiSpitfire.cpp
@@ -92,7 +92,7 @@ void vsiSpitfire::render()
 void vsiSpitfire::update()
 {
     // Check for position or size change
-    long *settings = globals.simVars->readSettings(name, xPos, yPos, size);
+    int *settings = globals.simVars->readSettings(name, xPos, yPos, size);
 
     xPos = settings[0];
     yPos = settings[1];
